scans buttons and dials now
move adc stuff to adcmanager, add dial helper
scope sampling is kinda wonky since it doesn't pause sampling to capture a waveform, but will fix that soon

+++ firmware/Core/Inc/scope.h
@@ -93,7 +93,7 @@ typedef struct {
 	SampleMode sampleMode;
 	DrawMode drawMode;
 
-	uint16_t triggerIndex;
+	int32_t triggerIndex;
 	int32_t minVoltage;
 	int32_t maxVoltage;
 } ScopeSettings;
